moved actions to enum type
