+++ types/actions.ts
@@ -1,0 +1,4 @@
+export enum ACTIONS {
+    ADD = "Add",
+    EDIT = "Edit",
+}
